perf: improve OpenAI and Gemini code
