+++ Easydict/Swift/Service/OpenAI/BaseOpenAIService.swift
@@ -49,7 +49,7 @@ public class BaseOpenAIService: LLMStreamService {
                     if let content = chatResult.choices.first?.delta.content {
                         resultText += content
                     }
-                    handleResult(queryType: queryType, resultText: resultText, error: nil, completion: completion)
+                    updateResultText(resultText, queryType: queryType, error: nil, completion: completion)
                 case let .failure(error):
                     // For stream requests, certain special cases may be normal for the first part of the data transfer, but the final parsing is incorrect.
                     var text: String?
@@ -61,12 +61,7 @@ public class BaseOpenAIService: LLMStreamService {
                         logError("\(name())-(\(model)) error: \(error.localizedDescription)")
                         logError(String(describing: error))
                     }
-                    handleResult(
-                        queryType: queryType,
-                        resultText: text,
-                        error: err,
-                        completion: completion
-                    )
+                    updateResultText(text, queryType: queryType, error: err, completion: completion)
                 }
             }
 
@@ -75,75 +70,17 @@ public class BaseOpenAIService: LLMStreamService {
 
             if !result.isStreamFinished {
                 if let error {
-                    handleResult(queryType: queryType, resultText: nil, error: error, completion: completion)
+                    updateResultText(nil, queryType: queryType, error: error, completion: completion)
                 } else {
                     // If already has error, we do not need to update it.
                     if result.error == nil {
                         resultText = getFinalResultText(text: resultText)
 //                        log("\(name())-(\(model)): \(resultText)")
-                        handleResult(queryType: queryType, resultText: resultText, error: nil, completion: completion)
+                        updateResultText(resultText, queryType: queryType, error: nil, completion: completion)
                         result.isStreamFinished = true
                     }
                 }
             }
         }
-    }
-}
-
-// MARK: OpenAI chat messages
-
-extension BaseOpenAIService {
-    typealias ChatCompletionMessageParam = ChatQuery.ChatCompletionMessageParam
-
-    func chatMessages(text: String, from: Language, to: Language) -> [ChatCompletionMessageParam] {
-        typealias Role = ChatCompletionMessageParam.Role
-
-        var chats: [ChatCompletionMessageParam] = []
-        let messages = translationMessages(text: text, from: from, to: to, systemPrompt: false)
-        for message in messages {
-            if let roleRawValue = message["role"],
-               let role = Role(rawValue: roleRawValue),
-               let content = message["content"] {
-                guard let chat = ChatCompletionMessageParam(role: role, content: content) else { return [] }
-                chats.append(chat)
-            }
-        }
-
-        return chats
-    }
-
-    func chatMessages(
-        queryType: EZQueryTextType,
-        text: String,
-        from: Language,
-        to: Language
-    )
-        -> [ChatCompletionMessageParam] {
-        typealias Role = ChatCompletionMessageParam.Role
-
-        var messages = [[String: String]]()
-
-        switch queryType {
-        case .sentence:
-            messages = sentenceMessages(sentence: text, from: from, to: to, systemPrompt: true)
-        case .dictionary:
-            messages = dictMessages(word: text, sourceLanguage: from, targetLanguage: to, systemPrompt: true)
-        case .translation:
-            fallthrough
-        default:
-            messages = translationMessages(text: text, from: from, to: to, systemPrompt: true)
-        }
-
-        var chats: [ChatCompletionMessageParam] = []
-        for message in messages {
-            if let roleRawValue = message["role"],
-               let role = Role(rawValue: roleRawValue),
-               let content = message["content"] {
-                guard let chat = ChatCompletionMessageParam(role: role, content: content) else { return [] }
-                chats.append(chat)
-            }
-        }
-
-        return chats
     }
 }
